+++ src/tls/client.h
@@ -12,7 +12,11 @@ namespace tls
     std::shared_ptr<Cert> cert;
 
   public:
-    Client(std::shared_ptr<Cert> cert_) : Context(true), cert(std::move(cert_))
+    Client(
+      std::shared_ptr<Cert> cert_,
+      const std::vector<std::string>& groups = {"P-521", "P-384", "P-256"}) :
+      Context(true, groups),
+      cert(std::move(cert_))
     {
       cert->configure_ssl(ssl, cfg);
     }